Refactor squire type and factory function
