+++ src/squire/types.ts
@@ -1,9 +1,10 @@
 import { Character } from "../character/types";
 import { Fighter } from "../fighter/types";
-import { NumberRange } from "../shared/types";
+
+export type BallismRange = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;
 
 export type Squire = Character & {
   servesTo: Fighter;
-  ballism: NumberRange;
-  phrase: "I'm a loser";
+  ballism: BallismRange;
+  speak: () => "I'm a loser";
 };
